Exploration of GWAS summary statistics
Code for exploration and analysis of how GWAS summary statistics file look like for both PTSD and Schizophrenia and what should be the considered p-value threshold for them when trying to chose the final SNP's to work with. We also looked into Odds Ratio, Effect Size and other associated parameters that were needed for our Polygenic risk score analysis.